We now create databases in model:Couchdb through direct http calls to Couchdb, rather than through Pouchdb.
This is because the other functions in this lib don't use Pouchdb. 
However, Pouchdb doesn't actually create a Couchdb database until it's 
used, thinking it is the only gateway to that database. That doesn't fit 
our purposes here. which is to call all sorts of user management 
functions that Pouchdb does not support.