Checking synchronization by following connections between users.
* Checking synchronization now occurs in the context of withDomeinFile.
* InvertedQueries now have a 'modifies' member when the perspective 
includes modifying verbs.
* The UserGraphProjection now contains an element 'modifiedUserRole' 
that has a Just EnumeratedRoleType value when the modification concerns 
a role with kind UserRole. We use this to avoid false positives by the 
synchronization checker.
* Instead of computing unconnected subgraphs, we now check whether there 
is a path in a UserGraphProjection from each modifying user to all 
others.
* The member 'onPropertyDelta' of Property is no longer an Array of 
InvertedQueries, but an object where the keys are EnumeratedRoleTypes 
and the values Array InvertedQuery. This allows us to treat properties 
used as AspectProperty in various roles as separate from each other.
* The UserGraph in the DomeinFile now is an Object instead of an Array 
of UserNodes.