Edge case: user role removal
When a user role is removed and that
user should know, we ran into a problem.
The transaction machinery looked up
the 'bottom' of the user role AFTER the transaction.
This meant after the actual removal of the role.
We now compile a map of user roles
and their bottoms while adding deltas.