Fast message acknowledgement
Prior to this version, we used to process
a transaction before acknowledging it to the broker.
While providing a safeguard in case the PDR was terminated
before it handled the transaction, this also caused
the broker to send the message multiple times
depending on how long it took to process it!