Fixing the optimalisation of finding a users role
We keep a tab on the role of the current user in a context and
likewise we mark a role that is ultimately
filled by the current user.
However, these annotations are purely local.
Thus, this fails for public contexts.
I've fixed this for the relevant functions by
making a case analysis.