Subtle error on setting properties
When a property type is represented twice on a role chain
(as can happen with Aspects), on setting a property
we must make sure we set it on the right role instance.
On checking whether a particular role instance already
has a value for that property, however, we ran through
the entire chain - so an 'upper' value is never set when
a 'lower' value is available.