Terminal slash for public url
The url at which a public role's resources are published,
results from a computation in the model.
However, this url should end on a slash to function properly in
the publishing functions.
It is not possible to check in compile time if this is the case.
In order to prevent obscure errors,
I now add a slash to the computation result
if it is missing.