Fixing database views
Since we no longer store instances immediately,
it may happen that we fail to retrieve an instance
from Couchdb even when it has just been created.
This only happens when we query a view, of course.
I've fixed it by adding to the result the filtered contents of the caches.