The 'serializationuser'
When we serialize a context, we do so from the perspective of a particular role. However, we should also provide a peer identity, in order for a transaction to be actually filled with deltas. If we use the systems own identifying user, no deltas are added to the transaction!
For that purpose we introduce the 'serializationuser': an instance of PerspectivesUsers that we use exclusively on serializing a context for export.