Optimised type checking + error fixed
We now compare types based on the DNF that is computed compile time. This speeds up several functions, as can be readily seen when compiling a model.

I also fixed a rather obscure error to do with transforming ADT with expandAndReduce. When a product has no terms after the transformation, it should be returned as Nothing instead of Just (PROD []).