Refactoring ADT normalisation
The direct cause was that normalisation appeared to fail for more complex types.
The semantics of SUM and PRODUCT have gotten exchanged. Fixed that and as a consequence, we now expand to Conjunctive Normal Form instead of Disjunctive Normal Form.
I have made the definition of CNF stricter than that of DNF.
Rewrote the normalisation.
In many cases, I had the specialisation functions backwards.
Anyhow, the semantics is now much more in line in how it is written.
All models now compile, even with the new perspective checking performed for (automatic) actions.